pyproject.toml
[project]
name = "test-qdrant-hybrid"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=3.5.0",
    "fastembed>=0.6.1",
    "numpy>=2.2.5",
    "openai>=1.93.0",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "qdrant-client>=1.13.3",
    "qdrant-hybrid-pipeline==0.2.2",
    "sentence-transformers>=4.1.0",
    "tqdm>=4.67.1",
]
